Add files via upload
The Economist Democracy Index is an annual index compiled by the UK-based magazine, The Economist,
through its Economist Intelligence Unit (EIU). This index rates the state of democracy across countries worldwide.

It is based on five categories:

Electoral Process and Pluralism:
This measures the fairness of elections, the freedom to form political parties, and the participation of various groups in the political process.

Functioning of Government: This evaluates how effectively governments operate, the quality of governance, the level of corruption, and whether political decisions are made in a transparent manner.

Political Participation: This assesses the level of public involvement in politics, including voter turnout, membership in political parties, and engagement with political issues.

Political Culture: This looks at the societal norms and attitudes towards politics, including the extent to which different viewpoints are tolerated and the relationship between government and religion.

Civil Liberties: This gauges the extent of individual freedoms, including freedom of speech, freedom of the press, and the right to protest.

Countries are classified into four types of regimes based on their scores:

Full democracies
Flawed democracies
Hybrid regimes
Authoritarian regimes
The Democracy Index is widely referenced in political and academic circles as a comprehensive measure of the state of democracy globally.
It provides a snapshot of the health of democracy in each country and also allows for comparisons across countries and over time.