Add files via upload
Imagine you have a big box of different LEGO bricks. Some are red, some are blue, some are big, and some are small. Your job is to group similar LEGO bricks together so that all the red ones go in one group, all the blue ones in another, and so on.

That’s what we are doing in this project—but instead of LEGO bricks, we have countries! 🌍

We want to group countries based on how democratic they are using their democracy index scores. Some countries have a lot of freedom (like full democracies), some have a little (like flawed democracies), and some have very little or no freedom (like authoritarian regimes).

To do this, we use clustering, which is like playing a sorting game! 🎯